+++ packages/valory/protocols/acn/__init__.py
@@ -23,8 +23,8 @@ This module contains the support resources for the acn protocol.
 It was created with protocol buffer compiler version `libprotoc 3.19.4` and aea version `1.6.0`.
 """
 
-from packages.fetchai.protocols.acn.message import AcnMessage
-from packages.fetchai.protocols.acn.serialization import AcnSerializer
+from packages.valory.protocols.acn.message import AcnMessage
+from packages.valory.protocols.acn.serialization import AcnSerializer
 
 
 AcnMessage.serializer = AcnSerializer
