+++ packages/valory/protocols/acn/message.py
@@ -27,8 +27,8 @@ from aea.configurations.base import PublicId
 from aea.exceptions import AEAEnforceError, enforce
 from aea.protocols.base import Message
 
-from packages.fetchai.protocols.acn.custom_types import AgentRecord as CustomAgentRecord
-from packages.fetchai.protocols.acn.custom_types import StatusBody as CustomStatusBody
+from packages.valory.protocols.acn.custom_types import AgentRecord as CustomAgentRecord
+from packages.valory.protocols.acn.custom_types import StatusBody as CustomStatusBody
 
 
 _default_logger = logging.getLogger("aea.packages.fetchai.protocols.acn.message")
