+++ tests/test_packages/test_connections/test_p2p_libp2p/test_communication.py
@@ -35,7 +35,7 @@ from aea.multiplexer import Multiplexer
 
 from packages.fetchai.protocols.default import DefaultSerializer
 from packages.fetchai.protocols.default.message import DefaultMessage
-from packages.open_aea.connections.p2p_libp2p.connection import NodeClient, Uri
+from packages.valory.connections.p2p_libp2p import NodeClient, Uri
 
 from tests.conftest import (
     _make_libp2p_connection,
