+++ tests/test_packages/test_connections/test_p2p_libp2p/test_fault_tolerance.py
@@ -31,7 +31,7 @@ from aea.multiplexer import Multiplexer
 
 from packages.fetchai.protocols.default.message import DefaultMessage
 from packages.fetchai.protocols.default.serialization import DefaultSerializer
-from packages.open_aea.connections.p2p_libp2p.check_dependencies import build_node
+from packages.valory.connections.p2p_libp2p import build_node
 
 from tests.common.utils import wait_for_condition
 from tests.conftest import (
