+++ tests/test_packages/test_connections/test_p2p_libp2p/test_public_dht.py
@@ -33,10 +33,10 @@ from aea.multiplexer import Multiplexer
 from aea.test_tools.test_cases import AEATestCaseMany
 
 from packages.fetchai.protocols.default.message import DefaultMessage
-from packages.open_aea.connections.p2p_libp2p.connection import (
+from packages.valory.connections.p2p_libp2p import (
     PUBLIC_ID as P2P_CONNECTION_PUBLIC_ID,
 )
-from packages.open_aea.connections.p2p_libp2p_client.connection import (
+from packages.valory.connections import (
     PUBLIC_ID as P2P_CLIENT_CONNECTION_PUBLIC_ID,
 )
