+++ tests/test_packages/test_connections/test_p2p_libp2p_client/test_aea_cli.py
@@ -26,7 +26,7 @@ from aea.helpers.base import CertRequest
 from aea.multiplexer import Multiplexer
 from aea.test_tools.test_cases import AEATestCaseEmpty
 
-from packages.open_aea.connections.p2p_libp2p_client.connection import PUBLIC_ID
+from packages.valory.connections import PUBLIC_ID
 
 from tests.conftest import (
     DEFAULT_LEDGER,
