+++ tests/test_packages/test_connections/test_p2p_libp2p_client/test_errors.py
@@ -34,7 +34,7 @@ from aea.helpers.base import CertRequest
 from aea.identity.base import Identity
 from aea.multiplexer import Multiplexer
 
-from packages.open_aea.connections.p2p_libp2p_client.connection import (
+from packages.valory.connections import (
     NodeClient,
     P2PLibp2pClientConnection,
     POR_DEFAULT_SERVICE_ID,
